microservices/user-service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Muhammad Zain Attiq <zainatteeq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlmodel = "^0.0.19"
psycopg = "^3.1.19"
aiokafka = "^0.10.0"
psycopg2 = "^2.9.9"
pytest = "^8.2.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
